Pre-boot the Ember app and fail early
If the Ember app fails to boot for whatever reason, it doesn't make any sense to
keep an express server around and return 500 on every single request. Instead,
we should pre-boot the Ember app *before* creating the express server and bail
if we encountered any errors.

Step 1/N to better separate the roles between the components.